requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
aiofiles>=23.0.0
pyfiglet>=1.0.2
colorama>=0.4.6
